+++ threedi_urban_eia_nl/process_results.py
@@ -279,7 +279,7 @@ def download_results(
     while len(remaining) > 0:
         for simulation in remaining:
             simulation_id: int = simulation[0]
-            isahw: str = simulation[1].split("isahw")[1]
+            simulation_name: str = simulation[1]
             printProgressBar(total - len(remaining), total, "Downloading result files")
             status: SimulationStatus = api_call(
                 api.simulations_status_list, simulation_id
@@ -319,7 +319,7 @@ def download_results(
 
                     if debug and result.filename.startswith("log"):
                         "Download log files and unzip"
-                        sim_dir = simulations_dir / f"{simulation_id}-isahw{isahw}"
+                        sim_dir = simulations_dir / f"{simulation_id}-{simulation_name}"
                         sim_dir.mkdir(parents=True)
                         download = api_call(
                             api.simulations_results_files_download,