+++ CHANGES.rst
@@ -2,10 +2,25 @@ Changelog of threedi-urban-eia-nl
 ===================================================
 
 
-0.0.6 (unreleased)
+0.0.8 (unreleased)
 ------------------
 
 - Nothing changed yet.
+
+
+0.0.7 (2026-02-17)
+------------------
+
+- Tiny change in a comment to re-attempt release after failed attempt to release 0.0.6
+
+
+0.0.6 (2026-02-17)
+------------------
+
+- Set aggregation settings instead of validating schematisation
+- Improve method for parsing datetime from filename
+- Add option to use existing saved states simulation and skip creating a new one
+- More flexible rain file naming requirements + documentation
 
 
 0.0.5 (2025-12-19)